Move Constants into new Internal namespace
