+++ test/Silverpop.Core.Tests/TransactMessageRecipientTests.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Silverpop.Core.Internal;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Xunit;
